app/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
duckduckgo-search>=7.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
faiss-cpu>=1.7.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
